+++ src/main/java/GDSCKNU/CitySavior/controller/ReportsController.java
@@ -6,9 +6,9 @@ import GDSCKNU.CitySavior.dto.report.response.MapReportsResponse;
 import GDSCKNU.CitySavior.dto.report.response.ReportDetailResponse;
 import GDSCKNU.CitySavior.dto.report.request.ReportRequest;
 import GDSCKNU.CitySavior.dto.report.response.StatisticsResponse;
-import GDSCKNU.CitySavior.service.AIService;
-import GDSCKNU.CitySavior.service.ReportService;
-import GDSCKNU.CitySavior.service.StorageService;
+import GDSCKNU.CitySavior.service.ai.AIService;
+import GDSCKNU.CitySavior.service.report.ReportService;
+import GDSCKNU.CitySavior.service.storage.StorageService;
 import jakarta.validation.Valid;
 import java.util.List;
 import java.util.Map;
