+++ src/main/java/GDSCKNU/CitySavior/converter/report/ReportDetailResponseConverter.java
@@ -1,10 +1,10 @@
-package GDSCKNU.CitySavior.converter;
+package GDSCKNU.CitySavior.converter.report;
 
 import GDSCKNU.CitySavior.dto.report.response.ReportDetailResponse;
 import GDSCKNU.CitySavior.entity.Report;
 import org.springframework.core.convert.converter.Converter;
 
-public class ReportToReportDetailResponseDtoConverter implements Converter<Report, ReportDetailResponse> {
+public class ReportDetailResponseConverter implements Converter<Report, ReportDetailResponse> {
     @Override
     public ReportDetailResponse convert(Report report) {
