+++ src/main/java/GDSCKNU/CitySavior/service/ai/AIService.java
@@ -1,6 +1,5 @@
-package GDSCKNU.CitySavior.service;
+package GDSCKNU.CitySavior.service.ai;
 
-import java.util.List;
 import org.springframework.web.multipart.MultipartFile;
 
 public interface AIService {
