+++ src/main/java/GDSCKNU/CitySavior/service/storage/StorageService.java
@@ -1,4 +1,4 @@
-package GDSCKNU.CitySavior.service;
+package GDSCKNU.CitySavior.service.storage;
 
 import java.io.IOException;
 import org.springframework.web.multipart.MultipartFile;
